Delade upp till olika filer för bättre struktur och användning
