+++ src/Components/ExportButton/ExportColumn.ts
@@ -1,0 +1,4 @@
+export interface ExportColumn<T extends object> {
+  field: keyof T;
+  headerName?: string;
+}
