+++ src/Components/index.ts
@@ -1,4 +1,3 @@
 export { default as EditableTable } from './EditableTable';
-export { default as ExportButton } from './ExportButton';
 export type { ExportColumn } from './ExportButton';
 export { default as Header } from './Header';
